fix: increased the perfomance by moving the logic from the MainPage to a custom useLocalStorage hook, moved static data to constants, added a favicon
