+++ src/pages/MainPage/index.tsx
@@ -1,101 +1,16 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import Column from '../../components/Column';
 import Header from '../../components/Header';
-import { theme } from '../../constants/styles/theme/theme';
+import initialColumns from '../../constants/initialColumns';
+import { useLocalStorage } from '../../hooks/useLocalStorage';
 import { ITask } from '../../interfaces/ITask';
-import getColorForTitle from '../../utils/getColorForTitle';
 import { StyledColumnList, StyledMain } from './index.styled';
 
-interface IColumn {
-  id: number;
-  title: string;
-  color: string;
-  tasks: ITask[];
-  isDefaultTitle: boolean;
-}
-
-function MainPage() {
-  const [columns, setColumns] = useState<IColumn[]>(() => {
-    const storedColumns = localStorage.getItem('kanban-columns');
-    return storedColumns
-      ? JSON.parse(storedColumns)
-      : [
-          { id: 1, title: 'To Do', color: theme.colors.MEDIUM_PRIORITY, tasks: [], isDefaultTitle: false },
-          { id: 2, title: 'In Progress', color: theme.colors.OK_PRIORITY, tasks: [], isDefaultTitle: false },
-          { id: 3, title: 'Done', color: theme.colors.LOW_PRIORITY, tasks: [], isDefaultTitle: false },
-        ];
-  });
+const MainPage: React.FC = () => {
+  const { columns, addColumn, removeColumn, updateColumnTitle, addTask, deleteTask, updateTask, moveTask } =
+    useLocalStorage(initialColumns);
   const [draggedTask, setDraggedTask] = useState<{ columnId: number; task: ITask } | null>(null);
-
-  useEffect(() => {
-    localStorage.setItem('kanban-columns', JSON.stringify(columns));
-  }, [columns]);
-
-  const addColumn = () => {
-    const newColumn: IColumn = {
-      id: Date.now(),
-      title: 'Column Title',
-      color: theme.colors.INITIAL_COLUMN_BACKGROUND,
-      tasks: [],
-      isDefaultTitle: true,
-    };
-    setColumns([...columns, newColumn]);
-  };
-
-  const removeColumn = (columnId: number) => {
-    setColumns((prevColumns) => prevColumns.filter((column) => column.id !== columnId));
-  };
-
-  const updateColumnTitle = (id: number, newTitle: string) => {
-    setColumns((prevColumns) =>
-      prevColumns.map((column) => {
-        if (column.id === id) {
-          const updatedColumn = { ...column, title: newTitle };
-          if (column.isDefaultTitle && newTitle !== 'Column Title') {
-            updatedColumn.color = getColorForTitle();
-            updatedColumn.isDefaultTitle = false;
-          }
-          return updatedColumn;
-        }
-        return column;
-      })
-    );
-  };
-
-  const addTask = (columnId: number, task: ITask) => {
-    setColumns(
-      columns.map((column) =>
-        column.id === columnId ? { ...column, tasks: [...column.tasks, { ...task, priority: task.priority || 'Priority' }] } : column
-      )
-    );
-  };
-
-  const deleteTask = (columnId: number, taskId: number) => {
-    setColumns(
-      columns.map((column) =>
-        column.id === columnId
-          ? {
-              ...column,
-              tasks: column.tasks.filter((task) => task.id !== taskId),
-            }
-          : column
-      )
-    );
-  };
-
-  const updateTask = (columnId: number, taskId: number, updatedTask: Partial<ITask>) => {
-    setColumns(
-      columns.map((column) =>
-        column.id === columnId
-          ? {
-              ...column,
-              tasks: column.tasks.map((task) => (task.id === taskId ? { ...task, ...updatedTask } : task)),
-            }
-          : column
-      )
-    );
-  };
 
   const handleDragStart = (columnId: number, task: ITask) => {
     setDraggedTask({ columnId, task });
@@ -106,19 +21,7 @@ function MainPage() {
       const { columnId: sourceColumnId, task } = draggedTask;
 
       if (sourceColumnId !== targetColumnId) {
-        setColumns((prevColumns) =>
-          prevColumns.map((column) => {
-            if (column.id === sourceColumnId) {
-              return { ...column, tasks: column.tasks.filter((t) => t.id !== task.id) };
-            }
-
-            if (column.id === targetColumnId) {
-              return { ...column, tasks: [...column.tasks, task] };
-            }
-
-            return column;
-          })
-        );
+        moveTask(sourceColumnId, targetColumnId, task);
       }
     }
     setDraggedTask(null);
@@ -152,6 +55,6 @@ function MainPage() {
       </StyledMain>
     </>
   );
-}
+};
 
 export default MainPage;
