+++ src/utils/generateUniqueId.ts
@@ -1,0 +1,7 @@
+let taskIdCounter = 1;
+
+const generateUniqueId = () => {
+  return taskIdCounter++;
+};
+
+export default generateUniqueId;
